+++ scientific/src/types/conversion_error.rs
@@ -27,4 +27,5 @@ impl Display for ConversionError {
 }
 
 #[cfg(feature = "std")]
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 impl std::error::Error for ConversionError {}