+++ scientific/src/types/error.rs
@@ -18,4 +18,5 @@ impl Display for Error {
 }
 
 #[cfg(feature = "std")]
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 impl std::error::Error for Error {}